+++ src/main/java/br/com/augustocosta/acs/integration/dto/dtoProdutoVenda.java
@@ -2,6 +2,8 @@ package br.com.augustocosta.acs.integration.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -11,7 +13,7 @@ public class dtoProdutoVenda {
     private String nome;
     private String marca;
     private String linha;
-    private Double preco;
+    private BigDecimal preco;
     private Integer quantidade;
 
     public dtoProdutoVenda() {
